% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan2_extract_summary.R
\name{MiXcan2_extract_summary}
\alias{MiXcan2_extract_summary}
\title{Extract the gene expression prediction summary}
\usage{
MiXcan2_extract_summary(model)
}
\arguments{
\item{model}{A direct output from MiXcan2() function, which includes the
prediction weights as well as other summaries of the prediction models.}
}
\value{
A row including summary statistics of the prediction model of the gene.
}
\description{
Extract the gene expression prediction summaries, such as  No of SNPs, correlation, R^2.
}
