% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan2_prediction.R
\name{MiXcan2_prediction}
\alias{MiXcan2_prediction}
\title{Predict the cell-type specific or non-specific expression levels of a gene with MiXcan model in new genetic data.}
\usage{
MiXcan2_prediction(weight, new_x)
}
\arguments{
\item{weight}{Weight matrix produced by MiXcan2_extract_weight()}

\item{new_x}{The genetic data used for prediction in a new data set.}
}
\value{
A N by 2 matrix indicating the predicted gene expression levels in two
cell types. If a non-specific model is used for prediction, the predicted values should be the same in two cell types.
}
\description{
Predict the cell-type specific or non-specific expression levels of a gene with MiXcan model in new genetic data.
}
