% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan2_Ensemble.R
\name{MiXcan2_ensemble}
\alias{MiXcan2_ensemble}
\title{MiXcan2 Ensemble}
\usage{
MiXcan2_ensemble(
  y,
  x,
  cov,
  pi,
  yName = NULL,
  xNameMatrix = NULL,
  B = 10,
  seed = NULL
)
}
\arguments{
\item{y}{The pre-cleaned expression level data for a single gene in N samples.}

\item{x}{A N by P matrix for all the genetic predictors used to predict the genetically regulated expression  of the gene.}

\item{cov}{A N by Q matrix for the covariates adjusted in the model (e.g. age, population stratification).}

\item{pi}{An estimation of cell-type faction of the cell type of interest.}

\item{yName}{Default is NULL. A row vector to save the expression information, such as gene ID, gene name.}

\item{xNameMatrix}{Default is NULL. A matrix to save the X matrix information,
such as variable ID, position, rsid, ref_allele, eff_allele.}

\item{B}{No. of ensemble models}

\item{seed}{seed}
}
\value{
A data frame with weight for cell 1 and 2, including the potential meta data for the SNP
}
\description{
MiXcan2 Ensemble
}
